+++ src/arch/mod.rs
@@ -21,6 +21,8 @@
     target_arch = "riscv64",
     target_arch = "mips",
     target_arch = "mips64",
+    target_arch = "powerpc",
+    target_arch = "powerpc64",
 )))]
 compile_error!("this target is not supported yet");
 
@@ -46,6 +48,8 @@ mod arm;
 mod armv8m;
 #[cfg(any(target_arch = "mips", target_arch = "mips64"))]
 mod mips;
+#[cfg(any(target_arch = "powerpc", target_arch = "powerpc64"))]
+mod powerpc;
 #[cfg(any(target_arch = "riscv32", target_arch = "riscv64"))]
 mod riscv;
 #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]