+++ src/lib.rs
@@ -18,10 +18,8 @@ Currently, x86, x86_64, ARM (v6-m, v7+), AArch64, RISC-V, MIPS32r2, MIPS64r2, Po
 | arm (v6-m, v7+)             | isize,usize,i8,u8,i16,u16,i32,u32                   | ✓              | ✓\[1]  |
 | arm (v7-a)                  | i64,u64                                             | ✓              | ✓      |
 | aarch64 \[2]                | isize,usize,i8,u8,i16,u16,i32,u32,i64,u64,i128,u128 | ✓              | ✓      |
-| riscv32                     | isize,usize,i32,u32                                 | ✓              | ✓\[1]  |
-| riscv32                     | i8,u8,i16,u16                                       | ✓              |        |
-| riscv64                     | isize,usize,i32,u32,i64,u64                         | ✓              | ✓\[1]  |
-| riscv64                     | i8,u8,i16,u16                                       | ✓              |        |
+| riscv32                     | isize,usize,i8,u8,i16,u16,i32,u32                   | ✓              | ✓\[1]  |
+| riscv64                     | isize,usize,i8,u8,i16,u16,i32,u32,i64,u64           | ✓              | ✓\[1]  |
 | mips \[3]                   | isize,usize,i32,u32                                 | ✓              | ✓      |
 | mips \[3]                   | i8,u8,i16,u16                                       | ✓              |        |
 | mips64 \[3]                 | isize,usize,i32,u32,i64,u64                         | ✓              | ✓      |