+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 - Support s390x.
 
+- Implement RISC-V 8-bit and 16-bit swap.
+
 ## [0.2.5] - 2022-05-30
 
 - Support ARMv6-M. (#4)