+++ CHANGELOG.md
@@ -16,6 +16,10 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 - Implement RISC-V 8-bit and 16-bit swap.
 
+- Implement MIPS32r2 8-bit and 16-bit swap.
+
+- Implement MIPS64r2 8-bit and 16-bit swap.
+
 ## [0.2.5] - 2022-05-30
 
 - Support ARMv6-M. (#4)